skills/gstdcoin/a2a/requirements.txt
requests
mcp
pydantic
tonsdk
pynacl
